build of maps on atendente
