+++ antendente/assets/data/orders.json
@@ -14,9 +14,9 @@
         "minDeliveryTime": 10,
         "maxDeliveryTime": 44,
         "rating": 1.5095123177204128,
-        "address": "983 Lewis Rapids",
-        "lat": 37.808,
-        "lng": -122.417743,
+        "address": "Rua do Mel 322",
+        "lat": -2.595058, 
+        "lng": -44.238981,
         "createdAt": "2022-04-16T12:03:33.225Z",
         "updatedAt": "2022-04-16T12:05:12.826Z"
       },
@@ -43,9 +43,9 @@
         "minDeliveryTime": 10,
         "maxDeliveryTime": 44,
         "rating": 1.5095123177204128,
-        "address": "983 Lewis Rapids",
-        "lat": 46.99190699528364,
-        "lng": 28.66937969418778,
+        "address": "Rua 20 de novembro",
+        "lat": -4.230102, 
+        "lng": -44.787755,
         "createdAt": "2022-04-16T12:03:33.225Z",
         "updatedAt": "2022-04-16T12:05:12.826Z"
       },
@@ -75,9 +75,9 @@
         "minDeliveryTime": 10,
         "maxDeliveryTime": 44,
         "rating": 1.5095123177204128,
-        "address": "983 Lewis Rapids",
-        "lat": 46.47190999528364,
-        "lng": 28.74937369418778,
+        "address": "Rua 12, casa 20",
+        "lat": -2.894993,
+        "lng": -44.880065,
         "createdAt": "2022-04-16T12:03:33.225Z",
         "updatedAt": "2022-04-16T12:05:12.826Z"
       },
@@ -101,15 +101,15 @@
       "_deleted": null,
       "Restaurant": {
         "id": "7d7965f7-1b33-4b8c-9bac-0c22bc905030",
-        "name": "Bambu Bar",
+        "name": "Bar o pelelexe",
         "image": "https://notjustdev-dummy.s3.us-east-2.amazonaws.com/uber-eats/restaurant3.jpeg",
         "deliveryFee": 0.039448871531492236,
         "minDeliveryTime": 10,
         "maxDeliveryTime": 44,
         "rating": 1.5095123177204128,
         "address": "983 Lewis Rapids",
-        "lat": 46.47190999528364,
-        "lng": 28.74937369418778,
+        "lat": -2.7700545026914045, 
+        "lng": -42.837127691884945,
         "createdAt": "2022-04-16T12:03:33.225Z",
         "updatedAt": "2022-04-16T12:05:12.826Z"
       },
@@ -133,15 +133,15 @@
       "_deleted": null,
       "Restaurant": {
         "id": "7d7965f7-1b33-4b8c-9bac-0c22bc905030",
-        "name": "Bambu Bar",
+        "name": "Tribos Bar",
         "image": "https://notjustdev-dummy.s3.us-east-2.amazonaws.com/uber-eats/restaurant3.jpeg",
         "deliveryFee": 0.039448871531492236,
         "minDeliveryTime": 10,
         "maxDeliveryTime": 44,
         "rating": 1.5095123177204128,
-        "address": "983 Lewis Rapids",
-        "lat": 46.47190999528364,
-        "lng": 28.74937369418778,
+        "address": "Rua Getúlio Vargas, 383",
+        "lat": -4.198135,
+        "lng": -45.404319,
         "createdAt": "2022-04-16T12:03:33.225Z",
         "updatedAt": "2022-04-16T12:05:12.826Z"
       },
@@ -165,15 +165,15 @@
       "_deleted": null,
       "Restaurant": {
         "id": "7d7965f7-1b33-4b8c-9bac-0c22bc905030",
-        "name": "Bambu Bar",
+        "name": "Bar da rosa",
         "image": "https://notjustdev-dummy.s3.us-east-2.amazonaws.com/uber-eats/restaurant3.jpeg",
         "deliveryFee": 0.039448871531492236,
         "minDeliveryTime": 10,
         "maxDeliveryTime": 44,
         "rating": 1.5095123177204128,
         "address": "983 Lewis Rapids",
-        "lat": 46.47190999528364,
-        "lng": 28.74937369418778,
+        "lat": -7.5301303833004365,
+        "lng":  -46.04256523528048,
         "createdAt": "2022-04-16T12:03:33.225Z",
         "updatedAt": "2022-04-16T12:05:12.826Z"
       },
